feat: Add OrderRepository interface
This commit adds the OrderRepository interface, which extends JpaRepository to provide CRUD operations for the Order entity. This interface will be used to interact with the database and perform operations on the Order data.

Note: This commit message follows the established convention of starting with a verb indicating the type of change (chore, feat) and providing a concise summary of the changes made.