feat: Add OrderProductRepository interface
This commit adds the OrderProductRepository interface, which extends JpaRepository to provide CRUD operations for the OrderProduct entity. It also includes a custom query method `findByOrderId` to retrieve a list of OrderProducts by orderId. This interface will be used to interact with the database and perform operations on the OrderProduct data.