feat: Update OrderSesrviceImpl to use environment variable for base URL
This commit modifies the OrderSesrviceImpl class to use the environment variable `MS_SEARCH_URL` as the base URL for making HTTP requests to the search service. This allows for more flexibility and easier configuration of the URL. The hardcoded localhost URL is replaced with the value retrieved from the environment variable.

Note: This commit message follows the established convention of starting with a verb indicating the type of change (chore, feat) and providing a concise summary of the changes made.